READ.ME: Three models, 2 Deep Learning and 1 Non Deep Learning
This project involved the implementation of various models for image analysis and pattern recognition using Python. The YOLO deep learning model was utilized for efficient and accurate object detection, while the RNN deep learning model was employed to analyze sequential patterns and improve recognition capabilities. Additionally, the ORB (Oriented FAST and Rotated BRIEF) model, a non-deep learning approach, was incorporated as an alternative to provide a balanced comparison of methodologies. Visualizations were created to enhance the interpretability of the results, and a comprehensive table was included to summarize and present the outcomes clearly.